ENH: Adds function to override credentials
Adds a function that gives credentials which are used in GbqConnector, instead of obtaining credentials in the default manner using web server based OAuth flow.

This allows the user to override the credentials in read_gbq and to_gbq, which is useful when there is an alternate method of obtaining credentials.  One example of this is when running on Google Compute Engine, where any process has access to the credentials associated with the account associated with the GCE instance.